requirements.txt
# === 정적 분석 ===
bandit>=1.7.7
semgrep>=1.50.0

# === LLM API 클라이언트 ===
openai>=1.12.0
google-genai>=1.0.0
anthropic>=0.18.0

# === API 서버 ===
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6

# === GitHub 연동 ===
PyGithub>=2.2.0
requests>=2.31.0

# === DB ===
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9

# === 코드 검증 / 테스트 ===
pytest>=8.0.0
httpx>=0.24.0

# === 비동기 작업 큐 ===
celery>=5.3.0
redis>=5.0.0

# === 민감정보 탐지 ===
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0

# === 암호화 ===
cryptography>=42.0.0

# === 유틸리티 ===
python-dotenv>=1.0.0
pyyaml>=6.0.1
